requirements.txt
aiogram>=3.24.0
python-dotenv>=1.0.0
aiofiles>=25.1.0
pydantic>=2.12.0
aiosqlite>=0.22.0

